django_ecommerce/ecommerce/requirements.txt
# Django
django==3.1.12

# Django REST Framework for API support
djangorestframework==3.14.0

# PostgreSQL support for Django (if using PostgreSQL)
psycopg2-binary==2.9.3

# MySQL/MariaDB support for Django (if using MySQL)
mysqlclient==2.1.0

# Other dependencies (like for MongoDB, Redis, etc., if used)
# djongo==1.3.6   # Uncomment if using MongoDB
# redis==4.0.2    # Uncomment if using Redis

# Other utilities
gunicorn==20.1.0  # WSGI server for production
